+++ EdenSynth/libeden/source/synth/envelope/EnvelopeFactory.cpp
@@ -10,7 +10,7 @@
 
 namespace eden::synth::envelope
 {
-	std::unique_ptr<Envelope> EnvelopeFactory::createEnvelope(double sampleRate, std::shared_ptr<EnvelopeParameters> parameters)
+	std::unique_ptr<Envelope> EnvelopeFactory::createEnvelope(float sampleRate, std::shared_ptr<EnvelopeParameters> parameters)
 	{
 		if (const auto adbdrParameters = std::dynamic_pointer_cast<ADBDRParameters>(parameters); adbdrParameters != nullptr)
 		{
@@ -24,12 +24,12 @@ namespace eden::synth::envelope
 		throw std::runtime_error("Invalid envelope parameters.");
 	}
 	
-	std::unique_ptr<Envelope> EnvelopeFactory::createADBDREnvelope(double sampleRate, std::shared_ptr<ADBDRParameters> parameters)
+	std::unique_ptr<Envelope> EnvelopeFactory::createADBDREnvelope(float sampleRate, std::shared_ptr<ADBDRParameters> parameters)
 	{
 		return std::make_unique<ADBDR>(sampleRate, *parameters);
 	}
 
-	std::unique_ptr<Envelope> EnvelopeFactory::createADSREnvelope(double sampleRate, std::shared_ptr<ADSRParameters> parameters)
+	std::unique_ptr<Envelope> EnvelopeFactory::createADSREnvelope(float sampleRate, std::shared_ptr<ADSRParameters> parameters)
 	{
 		return std::make_unique<ADSR>(sampleRate, *parameters);
 	}
