+++ EdenSynth/libeden/source/utility/TimeSampleConverter.cpp
@@ -8,22 +8,22 @@ namespace eden::utility
 {
 	constexpr unsigned TimeSampleConverter::MILLISECONDS_IN_SECOND;
 
-	unsigned TimeSampleConverter::timeToSamples(std::chrono::seconds duration, double sampleRate)
+	unsigned TimeSampleConverter::timeToSamples(std::chrono::seconds duration, float sampleRate)
 	{
 		return static_cast<unsigned>(duration.count() * sampleRate);
 	}
 
-	unsigned TimeSampleConverter::timeToSamples(std::chrono::milliseconds duration, double sampleRate)
+	unsigned TimeSampleConverter::timeToSamples(std::chrono::milliseconds duration, float sampleRate)
 	{
 		return static_cast<unsigned>(duration.count() * sampleRate / MILLISECONDS_IN_SECOND);
 	}
 
-	std::chrono::seconds TimeSampleConverter::samplesToSeconds(unsigned samples, double sampleRate)
+	std::chrono::seconds TimeSampleConverter::samplesToSeconds(unsigned samples, float sampleRate)
 	{
 		return std::chrono::seconds(static_cast<long long>(samples / sampleRate + 0.5));
 	}
 
-	std::chrono::milliseconds TimeSampleConverter::samplesToMilliseconds(unsigned samples, double sampleRate)
+	std::chrono::milliseconds TimeSampleConverter::samplesToMilliseconds(unsigned samples, float sampleRate)
 	{
 		return std::chrono::milliseconds(static_cast<long long>(MILLISECONDS_IN_SECOND * samples / sampleRate + 0.5));
 	}
